pyproject.toml
[project]
name = "lreata"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4==4.12.3",
    "braceexpand==0.1.7",
    "contourpy==1.2.0",
    "cycler==0.12.1",
    "einops==0.7.0",
    "fonttools==4.50.0",
    "fsspec==2024.3.1",
    "ftfy==6.2.0",
    "gdown==5.1.0",
    "geotorch==0.3.0",
    "huggingface-hub==0.21.4",
    "iopath==0.1.10",
    "kiwisolver==1.4.5",
    "matplotlib==3.8.3",
    "open-clip-torch==2.24.0",
    "opencv-python-headless==4.10.0.84",
    "packaging==24.0",
    "pandas==2.2.1",
    "portalocker==2.8.2",
    "protobuf==3.20.*",
    "pyparsing==3.1.2",
    "pysocks==1.7.1",
    "python-dateutil==2.9.0.post0",
    "pytz==2024.1",
    "regex==2023.12.25",
    "safetensors==0.4.2",
    "scikit-learn==1.5.1",
    "scipy==1.12.*",
    "seaborn==0.13.2",
    "six==1.16.0",
    "soupsieve==2.5",
    "tabulate==0.9.0",
    "tensorboard==2.17.0",
    "tensorboard-logger==0.1.0",
    "timm==0.9.16",
    "torchdiffeq==0.2.3",
    "tqdm==4.66.2",
    "tzdata==2024.1",
    "wcwidth==0.2.13",
    "webdataset==0.2.86",
    "yacs==0.1.8",
]
